powerpc/vas: Implement a simple FTW driver
The Fast Thread Wake-up (FTW) driver provides user space applications an
interface to the Core-to-Core functionality in POWER9. The driver provides
the device node/ioctl API to applications and uses the external interfaces
to the VAS driver to interact with the VAS hardware.

A follow-on patch provides detailed description of the API for the driver.

Changelog[]
	Implement VAS_FTW_SETUP ioctl
	Fold fixes from vas-ftw-v10 branch into this commit

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>