user-nx842: Add a selftest
Note that user space access to NX 842 engine is not supported. But,
this test provides an easy way to test the user space windows and VAS
fault handling using an NX engine. The test can be modified to use the
NX-GZIP engine when the kernel supports that. This test "reuses" the
FTW ioctl attribute structure for simplicity. When a user space nx-gzip
device driver is added, it could define VAS_TX_WIN_OPEN open ioctl
and this test should then switch over to that ioctl.

To be self-contained and avoid dependencies on other headers/include
dirs, the VAS/FTW headers are copied over to this test directory.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>