powerpc/vas: fix: delay clearing window instance
Haren was seeing a crash due to the winid got from pswid in a fault
from NX-GZIP engine. It appered to be dereferencing a null pointer,
possibly because the window was closed before NX completed processing
the request. Wonder why the check for credits during vas_win_close()
did not help/suffice.

Delay clearing the window instance until the hardware returns credits.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>